pickgok-ai-server/requirements.txt
flask>=2.3.0

faiss-cpu==1.13.0
numpy>=1.24.0
scikit-learn>=1.3.0

librosa==0.10.1

pandas>=2.0.0
tqdm==4.66.3

mysql-connector-python>=8.0.0

setuptools